move files to util and driver subdirs
