+++ app.py
@@ -3,8 +3,8 @@ import pandas as pd
 import biosnicar
 from biosnicar.rt_solvers.adding_doubling_solver import adding_doubling_solver
 from biosnicar.optical_properties.column_OPs import get_layer_OPs, mix_in_impurities
-from biosnicar.setup_snicar import setup_snicar
-from biosnicar.validate_inputs import validate_inputs
+from biosnicar.drivers.setup_snicar import setup_snicar
+from biosnicar.utils.validate_inputs import validate_inputs
 import plotly.express as px
 import streamlit as st
