+++ app_inputs.yaml
@@ -3,8 +3,8 @@ CTRL:
   WINDOW_SIZE: 9
   POLY_ORDER: 3
 
-
 RTM:
+  SOLVER: "AD"
   DIRECT: 1
   NBR_WVL: 480
   APRX_TYP: 2
