+++ biosnicar/drivers/get_albedo.py
@@ -3,11 +3,11 @@
 
 import numpy as np
 from pathlib import Path
-from biosnicar.validate_inputs import validate_inputs
+from biosnicar.utils.validate_inputs import validate_inputs
 from biosnicar.rt_solvers.adding_doubling_solver import adding_doubling_solver
 from biosnicar.optical_properties.column_OPs import get_layer_OPs, mix_in_impurities
-from biosnicar.display import display_out_data, plot_albedo
-from biosnicar.setup_snicar import setup_snicar
+from biosnicar.utils.display import display_out_data, plot_albedo
+from biosnicar.drivers.setup_snicar import setup_snicar
 from biosnicar.rt_solvers.toon_rt_solver import toon_solver
