+++ main.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 from pathlib import Path
-from biosnicar import get_albedo
+from biosnicar.drivers import get_albedo
 
 # call easy albedo func
 albedo = get_albedo.get("adding-doubling", plot=True, validate=True)
